Create thumbnail preview image creation lambda
This PR creates a new lambda function.  This function gets triggered by bitmovin notification when a finished encoding job completes transfering to S3.  We grab the job id and output location from the notification, create a new thumbnail for the encoding, and output it to the same location the encoding was sent to on S3.

Tags: bitmovin, images, thumbs, lambda