Use bitmovin-staging aws keys
After we create the thumbnail, we need to copy it from its location and write it to S3.  In order to do so, we need to use the aws keys for the bitmovin-staging user.  We will user the bitmovin-prod user for the production version of this lambda.

Tags: aws, s3, permissions